Feat: added bash and bat scripts to simplify running the backend
